pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "publicite-francais"
version = "0.1.0"
description = "French article scraper for vocabulary learning"
readme = "README.md"
authors = [{name = "roba7o"}]
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "requests>=2.25.1",
    "beautifulsoup4>=4.9.3",
    "lxml>=4.6.3",
    "urllib3>=1.26.5",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.23",
    "python-dotenv>=1.0.0",
    "tldextract>=5.0.0",
    "trafilatura>=1.8.0",
    "sqlparse>=0.4.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=8.2.1",
    "pytest-html>=4.1.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings  
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "W293",  # blank line contains whitespace
    "I001",  # import block is un-sorted or un-formatted
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"



